requirements.txt
opencv-python>=4.8.0
pillow>=10.2.0
streamlit>=1.29.0
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
langchain>=0.1.0
python-dotenv>=1.0.0